tests fiexd
